Fixed my bugs. Command acceptance is not built out.
I have so many questions on the flow of data. How do we intend on getting the users input from the broader community. What does that file look like, is there a reason to match to commands and then print off the commanders sheet? Will the detailer be getting the preference sheets before the commanders? Are the commanders going to use this app? IDK so many questions not enough time. The commanders preference sheet needs logic to add each command depending on how the introduction of data plays out